pyproject.toml
[project]
name = "wwm-gvg-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
]
